reddit/requirements.txt
django
djangorestframework>=3.9.0